Fixed all Warning
change sizeWithFont with sizeWithAttributes,change int with NSInterger,change Images and Titles Layout.

+++ popover/Demo/Calendar View/OCDaysView.m
@@ -53,11 +53,11 @@
 	NSDate *dateOnFirst = [calendar dateFromComponents:dateParts];
 	[dateParts release];
 	NSDateComponents *weekdayComponents = [calendar components:NSWeekdayCalendarUnit fromDate:dateOnFirst];
-	int weekdayOfFirst = [weekdayComponents weekday];	
+	NSInteger weekdayOfFirst = [weekdayComponents weekday];
     
     //NSLog(@"weekdayOfFirst:%d", weekdayOfFirst);
 
-	int numDaysInMonth = [calendar rangeOfUnit:NSDayCalendarUnit 
+	NSInteger numDaysInMonth = [calendar rangeOfUnit:NSDayCalendarUnit
 										inUnit:NSMonthCalendarUnit 
                                        forDate:dateOnFirst].length;
     
@@ -79,7 +79,7 @@
     
     [prevDateParts release];
     
-    int numDaysInPrevMonth = [calendar rangeOfUnit:NSDayCalendarUnit 
+    NSInteger numDaysInPrevMonth = [calendar rangeOfUnit:NSDayCalendarUnit
 										inUnit:NSMonthCalendarUnit 
                                        forDate:prevDateOnFirst].length;
     
@@ -87,9 +87,9 @@
     
     //Draw the text for each of those days.
     for(int i = 0; i <= weekdayOfFirst-2; i++) {
-        int day = numDaysInPrevMonth - weekdayOfFirst + 2 + i;
+        NSInteger day = numDaysInPrevMonth - weekdayOfFirst + 2 + i;
         
-        NSString *str = [NSString stringWithFormat:@"%d", day];
+        NSString *str = [NSString stringWithFormat:@"%ld", (long)day];
         
         
         
@@ -150,14 +150,14 @@
     [nextDateParts release];
     
     NSDateComponents *nextWeekdayComponents = [calendar components:NSWeekdayCalendarUnit fromDate:nextDateOnFirst];
-	int weekdayOfNextFirst = [nextWeekdayComponents weekday];
+	NSInteger weekdayOfNextFirst = [nextWeekdayComponents weekday];
     
     if(!endedOnSat) {
         //Draw the text for each of those days.
-        for(int i = weekdayOfNextFirst - 1; i < 7; i++) {
-            int day = i - weekdayOfNextFirst + 2;
+        for(NSInteger i = weekdayOfNextFirst - 1; i < 7; i++) {
+            NSInteger day = i - weekdayOfNextFirst + 2;
             
-            NSString *str = [NSString stringWithFormat:@"%d", day];
+            NSString *str = [NSString stringWithFormat:@"%ld", (long)day];
             
             CGContextSaveGState(context);
 //            CGContextSetShadowWithColor(context, shadow2Offset, shadow2BlurRadius, shadow2);
@@ -193,9 +193,9 @@
 	NSDate *dateOnFirst = [calendar dateFromComponents:dateParts];
 	[dateParts release];
 	NSDateComponents *weekdayComponents = [calendar components:NSWeekdayCalendarUnit fromDate:dateOnFirst];
-	int weekdayOfFirst = [weekdayComponents weekday];	
+	NSInteger weekdayOfFirst = [weekdayComponents weekday];
     
-	int numDaysInMonth = [calendar rangeOfUnit:NSDayCalendarUnit 
+	NSInteger numDaysInMonth = [calendar rangeOfUnit:NSDayCalendarUnit
 										inUnit:NSMonthCalendarUnit 
                                        forDate:dateOnFirst].length;
     didAddExtraRow = NO;
